% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_access_elements.R
\name{access_elements}
\alias{access_elements}
\alias{roc_get}
\alias{roc_get.roc_result_list}
\alias{roc_get.roc_results}
\alias{roc_get_all_results}
\alias{roc_cutoff}
\alias{roc_tp}
\alias{roc_fn}
\alias{roc_fp}
\alias{roc_tn}
\alias{roc_sens}
\alias{roc_spec}
\title{[!] Access elements of roc_result_list object}
\usage{
roc_get(obj, what)

\method{roc_get}{roc_result_list}(obj, what)

\method{roc_get}{roc_results}(obj, what)

roc_get_all_results(obj)

roc_cutoff(obj)

roc_tp(obj)

roc_fn(obj)

roc_fp(obj)

roc_tn(obj)

roc_sens(obj)

roc_spec(obj)
}
\arguments{
\item{obj}{Either an \code{roc_result_list} or an \code{roc_results} object.}

\item{what}{(\code{string(1)})\cr Name of element.}
}
\description{
[!!!] not tested yet.
}
\details{
Acces elements of \code{roc_result_list} and \code{roc_results} objects.
}
\seealso{
Other functions for ROC: 
\code{\link{calculate_performance}()},
\code{\link{print.as_str}()},
\code{\link{roc_performance_measures}},
\code{\link{roc_predict}()}
}
\author{
Vilmantas Gegzna
}
\concept{functions for ROC}
