% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-class_labels.R
\name{add_class_label}
\alias{add_class_label}
\alias{remove_class_label}
\title{Manage S3 class labels}
\usage{
add_class_label(x, new_class)

remove_class_label(x, drop_class)
}
\arguments{
\item{x}{An object to modify.}

\item{new_class}{(\code{character})\cr A name of a new class.
May be a vector of names.}

\item{drop_class}{(\code{character})\cr A name of a class to remove.
May be a vector of names.}
}
\description{
Add or remove S3 cpass labels.
}
\examples{
add_class_label(list("ok"), "ok_list")
}
