% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_performance.R
\name{calculate_performance}
\alias{calculate_performance}
\title{Performance measures for two-class classification}
\usage{
calculate_performance(truth, prediction, pos_label = levels(truth)[2])
}
\arguments{
\item{truth}{(\code{factor}) \cr A factor vector with "true"/reference
classes.}

\item{prediction}{A vector with predicted classes.}

\item{pos_label}{(\code{character(1)}) \cr A string with the name of
positive group. By default, the second level of factor
\code{truth} is treated as positive.}
}
\value{
A matrix with the following columns:
\itemize{
\item \code{tp} (number of true positives),
      \code{fn} (number of false negatives),
      \code{fp} (number of false positives),
      \code{tn} (number of true negatives),
      \code{sens} (sensitivity),
      \code{spec} (specificity),
      \code{ppv} (positive predictive value),
      \code{npv} (negative predictive value),
      \code{youden} (Youden's j index),
      \code{kappa} (Cohen's kappa).

\item
      attribute \code{labels} contains a named vector with elements
      \code{pos_label} and \code{neg_label}, which indicate labels of
       positive and negative groups respectively.
}\cr\cr
}
\description{
Performance measures for two-class classification
}
\examples{
(truth <- gl(n = 2, k = 3, length = 20, labels = c("H","S")))

(prediction <- rev(truth))


calculate_performance(truth, prediction)

calculate_performance(truth, prediction, pos_label = "S")

calculate_performance(truth, prediction, pos_label = "H")

}
\seealso{
Other functions for ROC: \code{\link{access_elements}},
  \code{\link{print.roc_df}},
  \code{\link{roc_performance_measures}},
  \code{\link{roc_predict}}
}
\author{
Vilmantas Gegzna
}
