% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_two_class_perormance.R
\name{calculate_two_class_perormance}
\alias{calculate_two_class_perormance}
\title{Performance measures for two-class classification}
\usage{
calculate_two_class_perormance(truth, prediction,
  pos_label = levels(truth)[2])
}
\arguments{
\item{truth}{(\code{factor}) \cr A factor vector with "true"/reference
classes.}

\item{prediction}{A vector with predicted classes.}

\item{pos_label}{(\code{character(1)}) \cr A string with the name of
positive group. By default, the second level of factor
\code{truth} is treated as positive.}
}
\value{
A data frame with the following results:
\itemize{
\item \bold{\code{$all_results}} is a data frame with columns
      \code{neg_label}, \code{pos_label} labels of negative and positive
      groups respectively,
      \code{tp} (number of true positives),
      \code{fn} (number of false negatives),
      \code{fp} (number of false positives),
      \code{tn} (number of true negatives),
      ... [!!!]

}\cr\cr
}
\description{
Performance measures for two-class classification
}
\examples{
(truth <- gl(n = 2, k = 3, length = 20, labels = c("H","S")))

(prediction <- rev(truth))


calculate_two_class_perormance(truth, prediction)

calculate_two_class_perormance(truth, prediction, pos_label = "S")

calculate_two_class_perormance(truth, prediction, pos_label = "H")

}
\seealso{
Other functions for ROC: \code{\link{access_elements}},
  \code{\link{print.roc_df}},
  \code{\link{roc_performance_measures}},
  \code{\link{roc_predict}}
}
\author{
Vilmantas Gegzna
}
