% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_class-labels.R
\name{class_add}
\alias{class_add}
\alias{class_remove}
\title{Manage S3 class labels}
\usage{
class_add(x, new_class)

class_remove(x, drop_class)
}
\arguments{
\item{x}{An object to modify.}

\item{new_class}{(\code{character})\cr A name of a new class.
May be a vector of names.}

\item{drop_class}{(\code{character})\cr A name of a class to remove.
May be a vector of names.}
}
\description{
Add or remove S3 cpass labels.
}
\examples{
class_add(list("ok"), "ok_list")

}
