% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo_.R
\name{cvo_get_info}
\alias{cvo_get_info}
\alias{cvo_get_inds.cvo_mlr}
\alias{cvo_get_inds.default}
\alias{cvo_get_sample_size}
\alias{cvo_get_seeds}
\alias{cvo_count_folds}
\title{Utility functions for cross vadidation (cvo) object}
\usage{
cvo_get_info(cvo)

\method{cvo_get_inds}{cvo_mlr}(cvo, fold, type = c("asis", "train", "test"))

\method{cvo_get_inds}{default}(cvo, fold, type = c("asis", "train", "test"))

cvo_get_sample_size(cvo)

cvo_get_seeds(cvo)

cvo_count_folds(cvo)
}
\description{
Function \code{cvo_get_info} returns information about the
cross-validation object \code{cvo}.

Function \code{cvo_get_sample_size} returns sample
 size of the object used to create the
cross-validation object \code{cvo}.

Function \code{cvo_get_seeds} returns seeds of (pseudo)random number
generator used for each repetition of splitting to folds.

Function \code{cvo_count_folds} returns total number of folds in a
 \code{cvo} object.
}
\seealso{
\code{\link{cvo_create_folds}}
}
