% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_values.R
\name{get_var_values}
\alias{get_var_values}
\title{Get vector of variable values}
\usage{
get_var_values(x, data)
}
\arguments{
\item{x}{Either a name (string) or a positive integer which indicates
position of variable in \code{data} or a vector.}

\item{data}{A data frame (or sililar structure).}
}
\value{
\itemize{
\item If \code{x} is a vector of at least 2 elements,
                  this vector is returned.
\item If \code{x} is a string or a number (length 1),
                  then \code{data[[x]]} is returned.
\item If \code{x} is \code{NULL}, \code{NULL} is returned.
}
}
\description{
Get vector of variable values
}
\examples{
library(multiROC)
dataset <- head(PlantGrowth)

get_var_values(x = "group", data = dataset)

get_var_values(x = c("group", "group"), data = dataset)

get_var_values(x = NULL, data = dataset)


}
