% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_manyROC_package.R
\docType{package}
\name{manyROC}
\alias{manyROC}
\alias{.onAttach}
\alias{manyROC-package}
\title{ManyROC -- tools for ROC analysis}
\usage{
.onAttach(libname, pkgname)
}
\description{
ManyROC -- tools for ROC analysis.
}
\author{
Vilmantas Gegzna
}
