% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_kappa.R
\name{measure_kappa}
\alias{measure_kappa}
\title{Cohen's kappa}
\usage{
measure_kappa(truth = NULL, response = NULL, conf_mat = NULL)
}
\arguments{
\item{truth}{a vector with true (reference) values.}

\item{response}{a vector with response (predicted) values.}

\item{conf_mat}{a table similar to (\code{table(truth, response, useNA = "no")}).}
}
\value{
[!!!]
}
\description{
This function is based on function `measureKAPPA` from `mlr` package.
}
\examples{
# [!!!] // none //

# matrix(c("TP", "FN", "FP", "TN"), 2)

make_conf_matrix <- function(TP, FN, FP, TN)
                 matrix(c(TP, FN, FP, TN), 2)

}
\seealso{
Other measures_: \code{\link{measure_wkappa}}
}
