% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp-parallelSetSeed.R
\name{parallelSetSeed}
\alias{parallelSetSeed}
\title{Set seeds for reproducible parallel computing with 'parallelMap' package}
\usage{
parallelSetSeed(iseed)
}
\arguments{
\item{iseed}{An integer to be supplied to \code{\link{set.seed}}, or
    \code{NULL} not to set reproducible seeds.}
}
\description{
A wrapper for \pkg{parallel}::\code{\link[parallel]{clusterSetRNGStream}}.
}
