% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_analysis.R, R/z print-methods.R
\name{roc_analysis}
\alias{roc_analysis}
\alias{print.roc_results}
\alias{print.roc_optimal}
\alias{print.roc_info}
\title{[!] Carry the ROC analysis}
\usage{
roc_analysis(x, gr, pos_label = levels(gr)[2], pos_is_larger = NULL,
  optimize_by = "bac", results = "all")

\method{print}{roc_results}(x, ...)

\method{print}{roc_optimal}(x, digits = 3, ...)

\method{print}{roc_info}(x, ...)
}
\arguments{
\item{x}{(\code{numeric}) \cr A numeric vector.\cr
(in \code{print} function) An object to print.}

\item{gr}{(\code{factor}) \cr A factor vector with two levels.}

\item{pos_label}{(\code{character(1)}) \cr A string with the name of
positive group.}

\item{pos_is_larger}{(\code{NULL}|\code{TRUE}|\code{FALSE}) \cr
A flag indicating, if values of positive group are (on average) larger.
If \code{NULL}, this option is determined basing on data using 10\%
trimmed mean.}

\item{optimize_by}{(\code{string(1)}) \cr [!!!] Method to determine the
optimal cut-off value.
Current options: \code{"bac"},
                \code{"youden"},
                \code{"kappa"}.}

\item{results}{(\code{character(1)}) \cr A string indicating which results
should be returned: either \code{"all"} (as described in
section "Values") or just \code{"optimal"}.}

\item{...}{[!!!] Passed to further methods.}

\item{digits}{(\code{integer(1)})\cr Number of significant digits to display.}
}
\value{
A list (which also inherits from class \code{"roc_result_list"})
with three fields: \code{$info}, \code{$optimal}, \code{$all_results}.\cr\cr

\bold{\code{$info}} is a data frame with columns
      \code{var_name} - empty string reserved for variable name,
      \code{neg_label}, \code{pos_label} labels of negative and positive
                        groups respectively,
      \code{n_neg}, \code{n_pos}, \code{n_total} - number of negative
and positive cases as well as number of cases in total.

\bold{\code{$optimal}} one row from \code{$all_results}, which was
determined as having optima threshold (cut-off) value. Sometimes it can be
several rows, if the performance is equally good. \cr\cr

\bold{\code{$all_results}} is a data frame with columns
       \code{cutoffs} for cutoff values,
       \code{tp} (number of true positives),
       \code{fn} (number of false negatives),
       \code{fp} (number of false positives),
       \code{tn} (number of true negatives),
       ... [!!!] \cr\cr
}
\description{
DO the ROC (receiver operating characteristic) analysis and calculate
vector of cut-off values and associated number of
true positives (TP),
false negatives (FN),
false positives (FP), and
true negatives (TN) as well as performance measures such as sensitivity,
specificity, etc.
\cr[!!!]
}
\note{
This function is based on functions \code{predict} and
\code{.compute.unnormalized.roc.curve}
from \pkg{ROCR} package.
}
\examples{
# Make some data
set.seed(1)
(x_ <- rnorm(10))
(gr_ <- gl(n = 2, k = 5, length = 10, labels = c("H","S")))

# Explore the functions

roc_analysis(x_, gr_)
roc_analysis(x_, gr_, pos_label = "H")

}
\seealso{
Other functions for ROC: \code{\link{access_elements}},
  \code{\link{roc_performance_measures}}
}
\author{
Vilmantas Gegzna
}
