% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R, R/roc_multiroc.R
\name{print.manyroc_result}
\alias{print.manyroc_result}
\alias{roc_manyroc}
\alias{roc_manyroc.matrix}
\alias{roc_manyroc.numeric}
\alias{roc_manyroc.data.frame}
\alias{roc_manyroc.hyperSpec}
\title{Carry out the manyROC analysis}
\usage{
\method{print}{manyroc_result}(x, ..., show_all = FALSE, perf_digits = 2,
  fmt = "\%.3g")

roc_manyroc(x, gr = NULL, optimize_by = "bac", ...)

\method{roc_manyroc}{matrix}(x, gr = NULL, optimize_by = "bac", ...,
  gr_sep = " vs. ")

\method{roc_manyroc}{numeric}(x, gr = NULL, optimize_by = "bac", ...)

\method{roc_manyroc}{data.frame}(x, gr = NULL, optimize_by = "bac", ...)

\method{roc_manyroc}{hyperSpec}(x, gr = NULL, optimize_by = "bac", ...)
}
\arguments{
\item{x}{A numeric matrix, a data frame, a \code{hyperSpec} object or another
type of object, convertible to a numeric matrix.}

\item{...}{Arguments passed for further methods.}

\item{show_all}{(\code{logical(1)})\cr A flag if whole dataset should be
printed. If \code{FALSE}, only a few fisrt and last rows
will be printed.}

\item{perf_digits}{(\code{integer(1)})\cr A number of decimals to display
for \emph{performance measures}. Default is 2.}

\item{fmt}{(\code{character(1)})\cr A string indicating number display
format for other numeric columns excluding performance measures.
The string will be passed to \code{\link[base]{sprintf}}.
Default is \code{"\%.3g"}.}

\item{gr}{Either a string (scalar, \code{character(1)}) or a \code{factor}
variable (a vector) which defines groups in \code{x}.}

\item{optimize_by}{(\code{string(1)})\cr A string with the name of
                classification performance measure to use. Currently
                available options:
\itemize{
      \item \code{"bac"} - for balanced accuracy (mean of sensitivity and
       specificity);\cr
      \item \code{"kappa"} - for Cohens kappa;\cr
      \item \code{"youden"} - for Youden's index;
}}

\item{gr_sep}{Group separator used to paste the names of groups. Default is
\code{" vs. "}.}
}
\value{
Object of classes \code{manyroc_result} and \code{data.frame} with
columns:
 \itemize{
      \item \code{compared_groups} Names of compared groups (separated by
                  \code{gr_sep} with default value \code{" vs. "});
      \item \code{feature} names of numeric features used in analysis;
      \item \code{median_neg} median value of negatives group;
      \item \code{cutoff} for optimal threshold/cut-off values and
                  corresponding performance measures;
      \item \code{median_pos} median value positives group;
      \item \code{TP} number of true positives;
      \item \code{FN} number of false negatives;
      \item \code{FP} number of false positives;
      \item \code{TN} number of true negatives;
      \item \code{sens} sensitivity (true positive rate, recall);
      \item \code{spec} specificity (true negative rate);
      \item \code{PPV} positive predictive value (precision);
      \item \code{NPV} negative predictive value;
      \item \code{BAC} balanced accuracy;
      \item \code{Youden} Youden’s J index;
      \item \code{Kappa} Cohen's kappa;
      \item \code{AUC} area under the ROC curve;
}
}
\description{
Do the ROC analysis (\code{\link{roc_analysis}}) for each column of \code{x}.
}
\details{
\enumerate{
\item Consider ordering of factor \code{gr} levels before the analysis, as the
first level will always be treated as negative and the last as positive.
\bold{E.g.}, if we have factor with 3 levels in this particular order
"A", "B", "C",  then "A" will always be negative, "C" always positive and
"B" positive, when compared to "A" and negative, when compared to "C".
The same principle applies if there are more than 3 levels.\cr
This is important determining what specificity and sensitivity, etc., means
in the context of group names: if positive is "A" then \emph{sensitivity}
will be related to group "A", but if "A" is negative, then \emph{specificity}
will be related to this group, and sensitivity to the other group.

\item For spectroscopic data: compare spectra of each pair of indicated groups
at each wavelength.
}
}
\examples{

library(manyROC)


# --- For numeric vectors objects ---

data(PlantGrowth)
roc_manyroc(PlantGrowth$weight, PlantGrowth$group)


# --- For dataframes objects ---

data(CO2)
roc_manyroc(CO2[, c("conc", "uptake")], CO2$Type)

data(OrchardSprays)
roc_manyroc(OrchardSprays$decrease, OrchardSprays$treatment)

# --- For hyperSpec objects ---

library(hyperSpec)
fluorescence

roc_manyroc(fluorescence[ , , 500~502], fluorescence$gr)

}
\author{
Vilmantas Gegzna
}
