% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_multiroc_cv.R
\name{roc_manyroc_cv}
\alias{roc_manyroc_cv}
\title{Carry out the manyROC analysis with cross-validation}
\usage{
roc_manyroc_cv(x, gr, optimize_by = "bac", cvo = cvo_create_folds(x, gr,
  seeds = seeds, kind = kind), seeds = NA_real_, kind = NULL)
}
\arguments{
\item{x}{A numeric matrix, a data frame, a \code{hyperSpec} object or another
type of object, convertible to a numeric matrix.}

\item{gr}{Either a string (scalar, \code{character(1)}) or a \code{factor}
variable (a vector) which defines groups in \code{x}.}

\item{optimize_by}{(\code{string(1)})\cr A string with the name of
                classification performance measure to use. Currently
                available options:
\itemize{
      \item \code{"bac"} - for balanced accuracy (mean of sensitivity and
       specificity);\cr
      \item \code{"kappa"} - for Cohens kappa;\cr
      \item \code{"youden"} - for Youden's index;
}}

\item{cvo}{a cross-validation object (cvo), created with function
\code{\link{cvo_create_folds}},
\pkg{caret} \code{\link[caret]{createFolds}}
or similar.}

\item{seeds}{(\code{NA_real_} | \code{NULL} | vector of integers)\cr
             Seeds for random number generator for each repetition.
  \itemize{
      \item If \code{seeds = NA_real_} (default), no seeds are set,
             parameter \code{kind} is also ignored.

      \item If \code{seeds = NULL} random seeds are generated
            automatically and registered in attribute \code{"seeds"}.

      \item If numeric vector, then these seeds will be used for each
            repetition of cross-validation.
            If the number of repetitions is greater than the number of
            provided seeds, additional seeds are generated and added to
            the vector. The first seed will be used to ensure
            reproducibility of the randomly generated seeds.
            }

             For more information about random number generation see
             \code{\link[base]{set.seed}}.}

\item{kind}{(\code{NULL} | \code{character})\cr
            The kind of (pseudo)random number generator. Default is
            \code{NULL}, which selects the currently-used generator
            (including that used in the previous session if the
            workspace has been restored): if no generator has been
            used it selects \code{"default"}.\cr

            Generator \code{"L'Ecuyer-CMRG"} is recommended if package
            \pkg{parallel} is used for for parallel computing.
            In this case each seed should have 6 elements neither the first
            three nor the last three should be all zero.
            More information at \code{\link[base]{set.seed}}.}
}
\value{
A data frame with results. The object also inferits from
 class \code{manyroc_cv} (for displaying purposes).
}
\description{
Do the ROC analysis (\code{\link{roc_analysis}}) with cross-validation
for each column of \code{x} (see also: \code{\link{roc_manyroc}}).
}
\examples{

library(manyROC)


# --- For numeric vectors objects ---

data(PlantGrowth)

set.seed(123456)

roc_manyroc_cv(PlantGrowth$weight, PlantGrowth$group)

roc_manyroc_cv(PlantGrowth$weight, gl(2, 1, 30))

}
\author{
Vilmantas Gegzna
}
