% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_merge_manyroc_cv_results.R
\name{roc_merge_manyroc_cv_results}
\alias{roc_merge_manyroc_cv_results}
\title{Merge 2 lists with manyROC CV results}
\usage{
roc_merge_manyroc_cv_results(x, y)
}
\arguments{
\item{x, y}{lists to merge}
}
\value{
A list wil results from \code{x} and \code{y}
}
\description{
Merge 2 lists with manyROC CV results
}
\examples{

# x <- readr::read_rds(path = "rez_366_a.RDS")
# y <- readr::read_rds(path = "rez_366_b.RDS")
#
# roc_merge_manyroc_cv_results(x, y)

# [!!!] unit test are needed.
}
