% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R, R/roc_multiroc.R
\name{print.multiroc_result}
\alias{print.multiroc_result}
\alias{roc_multiroc}
\alias{roc_multiroc.matrix}
\alias{roc_multiroc.numeric}
\alias{roc_multiroc.data.frame}
\alias{roc_multiroc.hyperSpec}
\title{[!] Carry out multiROC analysis}
\usage{
\method{print}{multiroc_result}(x, ..., digits = 2, cutoff_fmt = "\%.4g")

roc_multiroc(x, gr = NULL, optimize_by = "bac", ...)

\method{roc_multiroc}{matrix}(x, gr = NULL, optimize_by = "bac", ...,
  gr_sep = " vs. ")

\method{roc_multiroc}{numeric}(x, gr = NULL, optimize_by = "bac", ...)

\method{roc_multiroc}{data.frame}(x, gr = NULL, optimize_by = "bac", ...)

\method{roc_multiroc}{hyperSpec}(x, gr = NULL, optimize_by = "bac", ...)
}
\arguments{
\item{x}{A numeric matrix, a data frame, a \code{hyperSpec} object or another
type of object, convertible to a numeric matrix.}

\item{...}{Arguments passed for further methods.}

\item{digits}{(\code{integer(1)})\cr A number of decimals to display
for \emph{performance measures}. Default is 2.}

\item{cutoff_fmt}{(\code{character(1)})\cr A string with a number display
format for \emph{cut-off values}, \code{mean_neg}
 and \code{mean_pos}.
Will be passed to \code{\link[base]{sprintf}}. Default is
\code{"\%.4g"}.}

\item{gr}{Either a string (scalar, \code{character(1)}) or a \code{factor}
variable (a vector) which defines groups in \code{x}.}

\item{optimize_by}{(\code{string(1)})\cr A string with the name of
                classification performance measure to use. Currently
                available options:
\itemize{
      \item \code{"bac"} - for balanced accuracy (mean of sensitivity and specificity);\cr
      \item \code{"kappa"} - for Cohens kappa;\cr
      \item \code{"youden"} - for Youden's index;
}}

\item{gr_sep}{Group separator used to paste the names of groups. Default is
\code{" vs. "}.}
}
\value{
Object of classes \code{multiroc_result} and \code{data.frame} with
columns:#' \itemize{
      \item \code{compared_groups} Names of compared groups (separated by
                  \code{gr_sep} with default value \code{" vs. "});
      \item \code{feature} names of numeric features used in analysis;
      \item \code{mean_neg} 10\% trimmed mean of negatives;
      \item \code{cutoff} for optimal threshold/cut-off values and
                  corresponding performance measures;
      \item \code{mean_pos} 10\% trimmed mean of positives;
      \item \code{TP} number of true positives;
      \item \code{FN} number of false negatives;
      \item \code{FP} number of false positives;
      \item \code{TN} number of true negatives;
      \item \code{sens} sensitivity (true positive rate, recall);
      \item \code{spec} specificity (true negative rate);
      \item \code{PPV} positive predictive value (precision);
      \item \code{NPV} negative predictive value;
      \item \code{BAC} balanced accuracy;
      \item \code{Youden} Youden’s J index;
      \item \code{Kappa} Cohen's kappa;
      \item \code{AUC} area under the ROC curve;
}
}
\description{
For spectroscopic data: compare spectra of each pair of indicated groups
at each wavelength.
}
\examples{

library(multiROC)

# --- For numeric vectors objects ---

data(PlantGrowth)
roc_multiroc(PlantGrowth$weight, PlantGrowth$group)

# --- For dataframes objects ---

data(CO2)
roc_multiroc(CO2[, c("conc", "uptake")], CO2$Type)

# --- For hyperSpec objects ---

library(hyperSpec)
fluorescence
roc_multiroc(fluorescence[ , , 500~502], fluorescence$gr)
}
\author{
Vilmantas Gegzna
}
