% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_multiroc_cv.R
\name{roc_multiroc_cv}
\alias{roc_multiroc_cv}
\title{#' [!] Carry out the multiROC analysis with cross-validation}
\usage{
roc_multiroc_cv(x, gr, optimize_by = "bac", cvo = cvo_create_folds(x, gr,
  seeds = seeds, kind = kind), seeds = NULL, kind = "L'Ecuyer-CMRG")
}
\arguments{
\item{x}{A numeric matrix, a data frame, a \code{hyperSpec} object or another
type of object, convertible to a numeric matrix.}

\item{gr}{Either a string (scalar, \code{character(1)}) or a \code{factor}
variable (a vector) which defines groups in \code{x}.}

\item{optimize_by}{(\code{string(1)})\cr A string with the name of
                classification performance measure to use. Currently
                available options:
\itemize{
      \item \code{"bac"} - for balanced accuracy (mean of sensitivity and
       specificity);\cr
      \item \code{"kappa"} - for Cohens kappa;\cr
      \item \code{"youden"} - for Youden's index;
}}

\item{cvo}{a cross-validation object (cvo), created with function
\code{\link{cvo_create_folds}},
\pkg{caret} \code{\link[caret]{createFolds}}
or similar.}

\item{seeds}{(vector of integers | \code{NULL})\cr Seeds for random number
            generator for each repetition.\cr
            If \code{seeds = NULL} random seeds are generated.\cr
            If number of repetitions is
            greater than number of provided seeds, random seeds are
            generated and added to the provided ones. The first seed will
            be used to ensure reproducibility of the randomly generated
            seeds.\cr

             (See \code{\link[base]{set.seed}} for more information about
             random number generation).}

\item{kind}{(\code{character} | \code{NULL})\cr The kind of (pseudo)random
number generator. Default is \code{"L'Ecuyer-CMRG"} as it
provides the basis for the multiple streams used in package
\pkg{parallel}.
More information at \code{\link[base]{set.seed}}.}
}
\value{
A data frame with results. The object also inferits from
 class \code{multiroc_cv} (for displaying purposes).
}
\description{
#' [!] Carry out the multiROC analysis with cross-validation
}
\examples{

library(multiROC)


# --- For numeric vectors objects ---

data(PlantGrowth)

roc_multiroc_cv(PlantGrowth$weight, PlantGrowth$group)

roc_multiroc_cv(PlantGrowth$weight, gl(2, 1, 30))

}
