% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_performance.R
\name{roc_performance_measures}
\alias{roc_performance_measures}
\alias{calculate_acc}
\alias{roc_calculate_acc}
\alias{calculate_bac}
\alias{roc_calculate_bac}
\alias{calculate_auc}
\alias{roc_calculate_auc}
\alias{calculate_npv}
\alias{roc_calculate_npv}
\alias{calculate_ppv}
\alias{roc_calculate_ppv}
\alias{calculate_sensitivity}
\alias{roc_calculate_se}
\alias{calculate_specificity}
\alias{roc_calculate_sp}
\alias{calculate_youdens_j}
\alias{roc_youdens_j}
\alias{calculate_kappa}
\alias{roc_calculate_kappa}
\alias{calculate_wkappa}
\alias{roc_calculate_wkappa}
\title{[!!!] Calculate various performance measures for classificatory analysis}
\usage{
calculate_acc(TP, FN, FP, TN)

roc_calculate_acc(obj)

calculate_bac(SE, SP)

roc_calculate_bac(obj)

calculate_auc(SE, SP)

roc_calculate_auc(obj)

calculate_npv(TN, FN)

roc_calculate_npv(obj)

calculate_ppv(TP, FP)

roc_calculate_ppv(obj)

calculate_sensitivity(TP, FN)

roc_calculate_se(obj)

calculate_specificity(TN, FP)

roc_calculate_sp(obj)

calculate_youdens_j(SE, SP)

roc_youdens_j(obj)

calculate_kappa(TP, FN, FP, TN)

roc_calculate_kappa(obj)

calculate_wkappa(TP, FN, FP, TN)

roc_calculate_wkappa(obj)
}
\arguments{
\item{TP}{(\code{numeric}) Number of true positives.}

\item{FN}{(\code{numeric}) Number of false negatives.}

\item{FP}{(\code{numeric}) Number of false positives.}

\item{TN}{(\code{numeric}) Number of true negatives.}

\item{obj}{Either an \code{roc_result_list} or an \code{roc_results} object.}

\item{SE}{(\code{numeric}) Vector of sensitivities.}

\item{SP}{(\code{numeric}) Vector of specificities}
}
\description{
[!!!] The description is missing...
}
\seealso{
Other functions for ROC: \code{\link{access_elements}},
  \code{\link{roc_analysis}}
}
\author{
Vilmantas Gegzna
}
