% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_predict.R
\name{roc_predict}
\alias{roc_predict}
\title{Predict outcome for new data}
\usage{
roc_predict(object, x_new, ...)
}
\arguments{
\item{object}{Either an object which inherits from one of following classes:
 \code{roc_result_list},
 \code{roc_info},
 \code{roc_info_multi}, \bold{or} a data frame with exactly
 one row, which contains columns:
"cutoff" (numeric),
"below" (character),
"above" (character).}

\item{x_new}{A numeric vector with data to predict on.}

\item{...}{Arguments to further methods.}
}
\description{
Predict outcome for new data
}
\examples{
library(manyROC)

set.seed(1)
x <- rnorm(10)
gr <- gl(n = 2, k = 5, length = 10, labels = c("H","S"))
object <- roc_analysis(x, gr)

roc_predict(object, x_new = 1)

roc_predict(object, x_new = -1)

roc_predict(object, x_new = rnorm(20))

}
\seealso{
Other functions for ROC: \code{\link{access_elements}},
  \code{\link{calculate_performance}},
  \code{\link{print.as_str}},
  \code{\link{roc_performance_measures}}
}
