% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_predict_performance_by_gr.R
\name{roc_predict_performance_by_gr}
\alias{roc_predict_performance_by_gr}
\title{Calculate perfornamce for each pair of groups}
\usage{
roc_predict_performance_by_gr(object, x_new, gr_new)
}
\arguments{
\item{object}{Either an object of class \code{roc_info} or a data frame,
 which contains columns:
"cutoff" (numeric),
"below" (character),
"above" (character).}

\item{x_new}{A numeric vector to predict on.}

\item{gr_new}{A factor vector associated with `x_new`}
}
\value{
A dataframe with performance calculated on test data
}
\description{
This function works as combination of \code{\link{roc_predict}} and
\code{\link{calculate_performance}}, just \code{object} may have more than
1 row.
}
