% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_manyroc_with_cv.R
\name{sp_manyroc_with_cv}
\alias{sp_manyroc_with_cv}
\title{Do manyROC analysis with cross-validation for hyperSpec object}
\usage{
sp_manyroc_with_cv(Var, Spectra, k_folds = 5, times = 10, seeds = 2222222,
  kind = "L'Ecuyer-CMRG", n_min = k_folds)
}
\arguments{
\item{Var}{(\code{character(1)}\ \code{factor}) \cr
\bold{Either} the name of variable in \code{Spectra} which
contains the grouping variable
\bold{or} a factor vector (or convertible to factor) with
values for grouping.}

\item{Spectra}{\code{hyperSpec} object, that contains colum \code{ID} for
spectra blocking, columns with grouping variables as well
as spectroscopic information.}

\item{k_folds}{(positive \code{integer})  \cr
The number of folds in k-fold cross-validation.}

\item{times}{(positive \code{integer})  \cr
The number of repetitions in repeated k-fold
cross-validation.}

\item{seeds}{(\code{NULL} | \code{integer}) \cr
             Either a vector of integers of length \code{times} to set
             seed for each repetition of k-fold cross-validation.
             For more about seeds see \code{\link[base]{set.seed}()}.

             Each seed will be passed to \code{\link{cvo_create_folds}}.()}

\item{kind}{character or \code{NULL}.  If \code{kind} is a character
    string, set \R's RNG to the kind desired.  Use \code{"default"} to
    return to the \R default.  See \sQuote{Details} for the
    interpretation of \code{NULL}.}

\item{n_min}{\code{integer(1)} \cr
minimum acceptable number of unique samples per group.
Must be at least \code{k_folds} or bigger.}
}
\description{
[!!!] // No description yet //
}
\details{
Function \code{sp_manyroc_with_cv} will be \bold{renamed} in the future.
}
\examples{
library(manyROC)

fluorescence$ID  <- 1:nrow(fluorescence)
sp_manyroc_with_cv("gr", fluorescence[,,500~502], k_folds = 3, times = 2)

}
\seealso{
\code{\link{sp_manyroc_with_cv_by_variable}}
}
