% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo_get_inds.R, R/cvo_get_info.R
\name{cvo_get_inds}
\alias{cvo_get_inds}
\alias{cvo_get_inds.cvo_caret}
\alias{cvo_get_inds.cvo_mlr}
\alias{cvo_get_info}
\alias{cvo_get_sample_size}
\alias{cvo_get_seeds}
\alias{cvo_count_folds}
\alias{cvo_count_folds.cvo_caret}
\alias{cvo_count_folds.cvo_mlr}
\alias{cvo_get_fold_names}
\alias{cvo_get_fold_names.cvo_caret}
\alias{cvo_get_fold_names.cvo_mlr}
\title{Access information in a cvo object}
\usage{
cvo_get_inds(cvo, fold, type = c("asis", "train", "test"))

\method{cvo_get_inds}{cvo_caret}(cvo, fold, type = c("asis", "train", "test"))

\method{cvo_get_inds}{cvo_mlr}(cvo, fold, type = c("asis", "train", "test"))

cvo_get_info(cvo)

cvo_get_sample_size(cvo)

cvo_get_seeds(cvo)

cvo_count_folds(cvo)

\method{cvo_count_folds}{cvo_caret}(cvo)

\method{cvo_count_folds}{cvo_mlr}(cvo)

cvo_get_fold_names(cvo)

\method{cvo_get_fold_names}{cvo_caret}(cvo)

\method{cvo_get_fold_names}{cvo_mlr}(cvo)
}
\arguments{
\item{cvo}{a cross-validation object (\code{cvo}) created with
function \code{cvo_create_folds}.}

\item{fold}{either a name or an index of fold of interest.}

\item{type}{(string) one of options, indicating the kind of
            indices you are interested in:
\itemize{
     \item \code{"AsIs"} - indices as they are in the object,
     \item \code{"Train"} - indices of training set,
     \item \code{"Test"} - indices of test set.
}}
}
\description{
Utility functions for \code{cvo} object. More in section "Details."
}
\details{
Function \code{cvo_get_info} returns information about the
cross-validation object \code{cvo}.\cr

Function \code{cvo_get_sample_size} returns sample
size of the object used to create the cross-validation
object \code{cvo}.\cr

Function \code{cvo_count_folds} returns total number of folds in a
\code{cvo} object.\cr

Function \code{cvo_get_inds} extracts indices of indicated
set (either training or test) from cross-validation object (\code{cvo})
created with function \code{cvo_create_folds} when fold of interest is
indicated as \code{fold} (it can be either a name of fold or an index
of fold). If \code{cvo} does not contain indices of chosen
\code{type}, it returns a \bold{complement} to those indices and this
information is indicated in attributes of returned object.

Function \code{cvo_get_seeds} returns information about seeds of
(pseudo)random number generator used for each repetition of
splitting to folds.\cr
}
\examples{
cvo <- cvo_create_folds(fluorescence, folds = 10)
cvo

cvo_get_info(cvo)

cvo_get_seeds(cvo)

cvo_get_sample_size(cvo)

cvo_count_folds(cvo)

cvo_get_fold_names(cvo)

cvo_get_inds(cvo, 1)

cvo_get_inds(cvo, 1, "train")

cvo_get_inds(cvo, 1, "test")

}
\seealso{
\code{\link{cvo_create_folds}}
}
