% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_multiROC_package.R
\docType{package}
\name{manyROC}
\alias{manyROC}
\alias{manyROC-package}
\title{MultiROC algorithm for spectroscopic data analysis}
\description{
MultiROC algorithm for spectroscopic data analysis.
}
\author{
Vilmantas Gegzna
}
