% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_kappa.R
\name{measure_kappa}
\alias{measure_kappa}
\title{Cohen's kappa}
\usage{
measure_kappa(truth = NULL, response = NULL, conf_mat = NULL)
}
\arguments{
\item{truth}{a vector with true (reference) values.}

\item{response}{a vector with response (predicted) values.}

\item{conf_mat}{a table similar to (\code{table(truth, response, useNA = "no")}).}
}
\value{
[!!!]
}
\description{
This function is based on function `measureKAPPA` from `mlr` package.
}
\examples{

     truth <- rep(1:3, times = 50)
prediction <- rep(3:1, each = 50)

measure_kappa(truth, prediction)

square_matrix <- table(truth, prediction)
measure_kappa(conf_mat = square_matrix)


}
\seealso{
Other measures_: \code{\link{measure_wkappa}}
}
\concept{measures_}
