% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_extract_info.R
\name{roc_extract_info}
\alias{roc_extract_info}
\alias{roc_extract_info.multiroc_result}
\title{[!!!] Extract the main information necessary for prediction}
\usage{
roc_extract_info(obj, ...)

\method{roc_extract_info}{multiroc_result}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{pass to further methods}
}
\description{
[!!!] Extract the main information necessary for prediction
}
