% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R, R/roc_multiroc.R,
%   R/roc_multiroc_cv.R
\name{print.multiroc_result}
\alias{print.multiroc_result}
\alias{roc_multiroc}
\alias{roc_multiroc.matrix}
\alias{roc_multiroc.numeric}
\alias{roc_multiroc.data.frame}
\alias{roc_multiroc.hyperSpec}
\alias{roc_multiroc_cv}
\title{[!] Carry out the multiROC analysis}
\usage{
\method{print}{multiroc_result}(x, ..., show_all = FALSE, perf_digits = 2,
  fmt = "\%.3g")

roc_multiroc(x, gr = NULL, optimize_by = "bac", ...)

\method{roc_multiroc}{matrix}(x, gr = NULL, optimize_by = "bac", ...,
  gr_sep = " vs. ")

\method{roc_multiroc}{numeric}(x, gr = NULL, optimize_by = "bac", ...)

\method{roc_multiroc}{data.frame}(x, gr = NULL, optimize_by = "bac", ...)

\method{roc_multiroc}{hyperSpec}(x, gr = NULL, optimize_by = "bac", ...)

roc_multiroc_cv(x, gr, optimize_by = "bac", cvo = cvo_create_folds(x, gr,
  seeds), seeds = NULL)
}
\arguments{
\item{x}{A numeric matrix, a data frame, a \code{hyperSpec} object or another
type of object, convertible to a numeric matrix.}

\item{...}{Arguments passed for further methods.}

\item{show_all}{(\code{logical(1)})\cr A flag if whole dataset should be
printed. If \code{FALSE}, only a few fisrt and last rows
will be printed.}

\item{perf_digits}{(\code{integer(1)})\cr A number of decimals to display
for \emph{performance measures}. Default is 2.}

\item{fmt}{(\code{character(1)})\cr A string indicating number display
format for other numeric columns excluding performance measures.
The string will be passed to \code{\link[base]{sprintf}}.
Default is \code{"\%.3g"}.}

\item{gr}{Either a string (scalar, \code{character(1)}) or a \code{factor}
variable (a vector) which defines groups in \code{x}.}

\item{optimize_by}{(\code{string(1)})\cr A string with the name of
                classification performance measure to use. Currently
                available options:
\itemize{
      \item \code{"bac"} - for balanced accuracy (mean of sensitivity and specificity);\cr
      \item \code{"kappa"} - for Cohens kappa;\cr
      \item \code{"youden"} - for Youden's index;
}}

\item{gr_sep}{Group separator used to paste the names of groups. Default is
\code{" vs. "}.}

\item{cvo}{a cross-validation object (cvo), created with function
\code{\link{cvo_create_folds}},
\pkg{caret} \code{\link[caret]{createFolds}}
or similar.}

\item{seeds}{(vector of integers | \code{NULL})\cr Seeds for random number
            generator for each repetition.\cr
            If \code{seeds = NULL} random seeds are generated.\cr
            If number of repetitions is
            greater than number of provided seeds, random seeds are
            generated and added to the provided ones. The first seed will
            be used to ensure reproducibility of the randomly generated
            seeds.\cr

             (See \code{\link[base]{set.seed}} for more information about
             random number generation).}
}
\value{
Object of classes \code{multiroc_result} and \code{data.frame} with
columns:
 \itemize{
      \item \code{compared_groups} Names of compared groups (separated by
                  \code{gr_sep} with default value \code{" vs. "});
      \item \code{feature} names of numeric features used in analysis;
      \item \code{median_neg} median value of negatives group;
      \item \code{cutoff} for optimal threshold/cut-off values and
                  corresponding performance measures;
      \item \code{median_pos} median value positives group;
      \item \code{TP} number of true positives;
      \item \code{FN} number of false negatives;
      \item \code{FP} number of false positives;
      \item \code{TN} number of true negatives;
      \item \code{sens} sensitivity (true positive rate, recall);
      \item \code{spec} specificity (true negative rate);
      \item \code{PPV} positive predictive value (precision);
      \item \code{NPV} negative predictive value;
      \item \code{BAC} balanced accuracy;
      \item \code{Youden} Youden’s J index;
      \item \code{Kappa} Cohen's kappa;
      \item \code{AUC} area under the ROC curve;
}

A data frame with results. The object also inferits from
 class \code{multiroc_cv} (for displaying purposes).
}
\description{
For spectroscopic data: compare spectra of each pair of indicated groups
at each wavelength.
}
\details{
Consider ordering of factor \code{gr} levels before the analysis, as the
first level will always be treated as negative and the last as positive.\cr
\bold{E.g.}, if we have factor with 3 levels in this particular order
"A", "B", "C",  then "A" will always be negative, "C" always positive and
"B" positive, when compared to "A" and negative, when compared to "C".
The same principle applies if there are more than 3 levels.\cr
This is important determining what specificity and sensitivity, etc., means
in the context of group names: if positive is "A" then \emph{sensitivity}
will be related to group "A", but if "A" is negative, then \emph{specificity}
will be related to this group, and sensitivity to the other group.
}
\examples{

library(multiROC)


# --- For numeric vectors objects ---

data(PlantGrowth)
roc_multiroc(PlantGrowth$weight, PlantGrowth$group)


# --- For dataframes objects ---

data(CO2)
roc_multiroc(CO2[, c("conc", "uptake")], CO2$Type)

data(OrchardSprays)
roc_multiroc(OrchardSprays$decrease, OrchardSprays$treatment)

# --- For hyperSpec objects ---

library(hyperSpec)
fluorescence

roc_multiroc(fluorescence[ , , 500~502], fluorescence$gr)

\dontrun{\donttest{
rez <- roc_multiroc_cv(x = Spectra2, gr = "gr")

rez
}}

}
\author{
Vilmantas Gegzna
}
