% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_manyroc_with_cv.R,
%   R/sp_manyroc_with_cv_by_variable.R
\name{sp_manyroc_with_cv}
\alias{sp_manyroc_with_cv}
\alias{sp_manyroc_with_cv_by_variable}
\title{Do manyROC anamysis with cross-validation for hyperSpec object}
\usage{
sp_manyroc_with_cv(Var, Spectra, k_folds = 5, times = 10, seeds = 2222222,
  n_min = k_folds)

sp_manyroc_with_cv_by_variable(Spectra, variables_to_analyze, k_folds = 3,
  times = 10, seeds = 2222222)
}
\arguments{
\item{Var}{(\code{character(1)}\ \code{factor}) \cr
\bold{Either} the name of variable in \code{Spectra} which
contains the grouping variable
\bold{or} a factor vector (or convertible to factor) with
values for grouping.}

\item{Spectra}{\code{hyperSpec} object, that contains colum \code{ID} for
spectra blocking, columns with grouping variables as well
as spectroscopic information.}

\item{k_folds}{(positive \code{integer})  \cr
The number of folds in k-fold cross-validation.}

\item{times}{(positive \code{integer})  \cr
The number of repetitions in repeated k-fold cross-validation.}

\item{seeds}{(\code{NULL} | \code{integer}) \cr
Either a vector of integers of length \code{times} to set
seed for each repetition of k-fold cross-validation.
The seeds nust be at least of length 6 and meet other
requirements for \code{"L'Ecuyer-CMRG"} random number
generator.}

\item{n_min}{\code{integer(1)} \cr
minimum acceptable number of unique samples per group.
Must be at least \code{k_folds} or bigger.}

\item{variables_to_analyze}{(\code{character}) \cr
A character vector with names of grouping
variables to use in analysis.}
}
\value{
A list with results \cr
       [!!!] Description needs more specification.
}
\description{
[!!!] // No description yet //
}
\details{
Function \code{sp_manyroc_with_cv} will be \bold{renamed} in the future.
}
\examples{
library(manyROC)

fluorescence$ID  <- 1:nrow(fluorescence)
sp_manyroc_with_cv("gr", fluorescence[,,500~502], k_folds = 3, times = 2)


library(manyROC)

fluorescence$ID  <- 1:nrow(fluorescence)
sp_manyroc_with_cv_by_variable(fluorescence[,,500~501], c("gr", "class"),
                          k_folds = 3, times = 2)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

\dontrun{
\donttest{
 sp_manyroc_with_cv_by_variable(fluorescence, c("gr")
 }}
}
